v2_project/backend/requirements.txt
# Core API
fastapi>=0.104.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# AI & RAG (Haystack + PgVector)
haystack-ai>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
ollama>=0.1.0

# Database & Search
# elasticsearch removed
# qdrant removed
# pocketbase removed

# Utilities
httpx>=0.26.0
edge-tts>=6.1.9
faster-whisper>=0.10.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pandas>=2.1.0  # For CSV/Excel processing
openpyxl>=3.1.2 # For Excel files
pdfplumber>=0.10.0 # For PDF extraction
Pillow>=10.0.0 # For Image processing

# Automation & Scrapers
playwright>=1.40.0
beautifulsoup4>=4.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
